% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cusumchart.R
\name{cusumchart}
\alias{cusumchart}
\title{Cusum status charts}
\usage{
cusumchart(y, target, sdev, shift = NULL, k = 0.5, h = 5,
  headstart = FALSE, ylim, main, cex = 0.8)
}
\arguments{
\item{y}{Numeric vector of measures to plot. Mandatory.}

\item{target}{Target value of the process mean.}

\item{sdev}{Standard deviation of the process.}
}
\value{
Nothing.
}
\description{
Cusum status charts are a graphical display of the tabular cusum and may be
used to monitor the mean of a process. They show the one-sided upper and
lower cusums.
}
\details{
Lipsum.
}
\references{
Douglas C. Montgomery (2009). Introduction to Statistical Process
Control, Sixth Edition, John Wiley & Sons.
}

