% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qic.R
\name{plot.qic}
\alias{plot.qic}
\title{Plot qic object}
\usage{
\method{plot}{qic}(x, y = NULL, ...)
}
\arguments{
\item{x}{List object returned from the qic() function.}

\item{y}{Ignored. Included for compatibility with generic plot function.}

\item{...}{Further arguments to plot function.}
}
\value{
Creates a qic plot.
}
\description{
Plot qic object
}
\examples{
y <- rnorm(24)
p <- qic(y, plot.chart = FALSE)
plot(p)

}
